Move most maps-related files to :feature:map
