+++ app/src/fdroid/java/com/geeksville/mesh/ui/map/MapView.kt
@@ -72,9 +72,6 @@ import com.geeksville.mesh.copy
 import com.geeksville.mesh.database.entity.Packet
 import com.geeksville.mesh.model.Node
 import com.geeksville.mesh.model.UIViewModel
-import com.geeksville.mesh.model.map.CustomTileSource
-import com.geeksville.mesh.model.map.MarkerWithLabel
-import com.geeksville.mesh.model.map.clustering.RadiusMarkerClusterer
 import com.geeksville.mesh.ui.map.components.CacheLayout
 import com.geeksville.mesh.ui.map.components.DownloadButton
 import com.geeksville.mesh.ui.map.components.EditWaypointDialog
@@ -84,11 +81,14 @@ import com.geeksville.mesh.util.addCopyright
 import com.geeksville.mesh.util.addScaleBarOverlay
 import com.geeksville.mesh.util.createLatLongGrid
 import com.geeksville.mesh.util.formatAgo
-import com.geeksville.mesh.util.zoomIn
 import com.geeksville.mesh.waypoint
 import com.google.accompanist.permissions.ExperimentalPermissionsApi // Added for Accompanist
 import com.google.accompanist.permissions.rememberMultiplePermissionsState // Added for Accompanist
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import org.meshtastic.feature.map.cluster.RadiusMarkerClusterer
+import org.meshtastic.feature.map.model.CustomTileSource
+import org.meshtastic.feature.map.model.MarkerWithLabel
+import org.meshtastic.feature.map.zoomIn
 import org.osmdroid.bonuspack.utils.BonusPackHelper.getBitmapFromVectorDrawable
 import org.osmdroid.config.Configuration
 import org.osmdroid.events.MapEventsReceiver
