+++ feature/map/src/fdroid/java/org/meshtastic/feature/map/cluster/MarkerClusterer.java
@@ -15,14 +15,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.model.map.clustering;
+package org.meshtastic.feature.map.cluster;
 
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.Point;
 import android.view.MotionEvent;
 
-import com.geeksville.mesh.model.map.MarkerWithLabel;
+import org.meshtastic.feature.map.model.MarkerWithLabel;
 
 import org.osmdroid.util.BoundingBox;
 import org.osmdroid.views.MapView;
