+++ feature/map/src/fdroid/java/org/meshtastic/feature/map/cluster/RadiusMarkerClusterer.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.model.map.clustering;
+package org.meshtastic.feature.map.cluster;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -27,7 +27,7 @@ import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.view.MotionEvent;
 
-import com.geeksville.mesh.model.map.MarkerWithLabel;
+import org.meshtastic.feature.map.model.MarkerWithLabel;
 
 import org.osmdroid.bonuspack.R;
 import org.osmdroid.util.BoundingBox;
