+++ feature/map/src/fdroid/kotlin/org/meshtastic/feature/map/model/CustomTileSource.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.model.map
+package org.meshtastic.feature.map.model
 
 import org.osmdroid.tileprovider.tilesource.ITileSource
 import org.osmdroid.tileprovider.tilesource.OnlineTileSourceBase
